hft_trading_bot/requirements.txt
# ============================================================
#  HFT Trading Bot — Production Dependencies
# ============================================================
#  Install: pip install -r requirements.txt
# ============================================================

# Async HTTP & WebSocket
aiohttp>=3.9.0

# ONNX Runtime (CPU-only for minimal footprint)
onnxruntime>=1.17.0

# Numerical computation (NO Pandas in live pipeline)
numpy>=1.26.0

# Environment variable management
python-dotenv>=1.0.0

# Serialization (for scaler loading)
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
